+++ CHANGELOG.md
@@ -15,6 +15,18 @@ Types of changes:
 - `Security` in case of vulnerabilities.
 -->
 
+## [Unreleased]
+
+### Added
+
+- Define also `Base.cconvert`. Fixes [#10].
+
+### Fixed
+
+- Error while calling `pointer` ([#10]).
+
+[#10]: https://github.com/FedericoStra/ProtectedArrays.jl/issues/10
+
 ## [0.2.0]
 
 ### Added
@@ -51,6 +63,7 @@ Types of changes:
 
 Empty project.
 
+[Unreleased]: https://github.com/FedericoStra/ProtectedArrays.jl/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/FedericoStra/ProtectedArrays.jl/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/FedericoStra/ProtectedArrays.jl/compare/v0.0.0...v0.1.0
 [0.0.0]: https://github.com/FedericoStra/ProtectedArrays.jl/releases/tag/v0.0.0